+++ src/RoBMA.cc
@@ -6,7 +6,9 @@
 #include "distributions/DWN.h"
 #include "distributions/DWWN1.h"
 #include "distributions/DWWN2.h"
-//#include "distributions/DMN.h"
+#include "distributions/DMN.h"
+#include "distributions/DMNMIX.h"
+#include "distributions/DMNMIXCD.h"
 #include "distributions/DWMN1.h"
 #include "distributions/DWMN2.h"
 #include "distributions/DMNv.h"
@@ -43,7 +45,9 @@ namespace jags {
       insert(new DWN);
       insert(new DWWN1);
       insert(new DWWN2);
-      //insert(new DMN);
+      insert(new DMN);
+      insert(new DMNMIX);
+      insert(new DMNMIXCD);
       insert(new DWMN1);
       insert(new DWMN2);
       insert(new DMNv);